Fix tests for test runner executable
