+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/testSupport/commands/forceLwcTestDebugAction.test.ts
@@ -6,6 +6,7 @@
  */
 
 import { expect } from 'chai';
+import * as fs from 'fs';
 import * as path from 'path';
 import { assert, SinonStub, stub } from 'sinon';
 import * as uuid from 'uuid';
@@ -20,7 +21,7 @@ import {
   handleDidStartDebugSession,
   handleDidTerminateDebugSession
 } from '../../../../src/testSupport/commands/forceLwcTestDebugAction';
-import * as lwcTestRunner from '../../../../src/testSupport/testRunner';
+import * as lwcTestWorkspace from '../../../../src/testSupport/workspace';
 import {
   TestCaseInfo,
   TestInfoKind,
@@ -49,7 +50,7 @@ describe('Force LWC Test Debug - Code Action', () => {
     ],
     Thenable<any>
   >;
-  let lwcTestRunnerStub: SinonStub<[string], string | undefined>;
+  let getLwcTestRunnerExecutableStub: SinonStub<[string], fs.PathLike>;
   let processHrtimeStub: SinonStub<
     [([number, number] | undefined)?],
     [number, number]
@@ -64,7 +65,10 @@ describe('Force LWC Test Debug - Code Action', () => {
     debugStub = stub(vscode.debug, 'startDebugging');
     processHrtimeStub = stub(process, 'hrtime');
     telemetryStub = stub(telemetryService, 'sendCommandEvent');
-    lwcTestRunnerStub = stub(lwcTestRunner, 'getLwcTestRunnerExecutable');
+    getLwcTestRunnerExecutableStub = stub(
+      lwcTestWorkspace,
+      'getLwcTestRunnerExecutable'
+    );
     uuidStub.returns(mockUuid);
     telemetryStub.returns(Promise.resolve());
     debugStub.returns(Promise.resolve());
@@ -75,7 +79,7 @@ describe('Force LWC Test Debug - Code Action', () => {
     debugStub.restore();
     processHrtimeStub.restore();
     telemetryStub.restore();
-    lwcTestRunnerStub.restore();
+    getLwcTestRunnerExecutableStub.restore();
   });
 
   const root = /^win32/.test(process.platform) ? 'C:\\' : '/var';
@@ -136,7 +140,7 @@ describe('Force LWC Test Debug - Code Action', () => {
 
   describe('Debug Test Case', () => {
     it('Should send telemetry for debug test case', async () => {
-      lwcTestRunnerStub.returns(lwcTestExecutablePath);
+      getLwcTestRunnerExecutableStub.returns(lwcTestExecutablePath);
       const mockExecutionTime: [number, number] = [123, 456];
       processHrtimeStub.returns(mockExecutionTime);
       const debugConfiguration = getDebugConfiguration(command, args, cwd);
@@ -164,7 +168,7 @@ describe('Force LWC Test Debug - Code Action', () => {
 
   describe('Debug Test File', () => {
     beforeEach(() => {
-      lwcTestRunnerStub.returns(lwcTestExecutablePath);
+      getLwcTestRunnerExecutableStub.returns(lwcTestExecutablePath);
       mockTestResultWatcher();
     });
     afterEach(() => {
