+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/testSupport/testRunner/index.test.ts
@@ -6,7 +6,6 @@
  */
 import { TestRunner as UtilsTestRunner } from '@salesforce/salesforcedx-utils-vscode/out/src/cli/';
 import { expect } from 'chai';
-import * as fs from 'fs';
 import * as path from 'path';
 import { assert, SinonStub, stub, SinonSpy, spy } from 'sinon';
 import * as uuid from 'uuid';
@@ -19,68 +18,13 @@ import {
   TestType
 } from '../../../../src/testSupport/types';
 
-import { nls } from '../../../../src/messages';
-import { telemetryService } from '../../../../src/telemetry';
 import {
-  getLwcTestRunnerExecutable,
   TestRunner,
   TestRunType
 } from '../../../../src/testSupport/testRunner';
 import { InputBuffer } from 'uuid/interfaces';
 
 describe('LWC Test Runner', () => {
-  describe('getLwcTestRunnerExecutable Unit Tests', () => {
-    let existsSyncStub: SinonStub<[fs.PathLike], boolean>;
-    let notificationStub: SinonSpy<
-      [string, vscode.MessageOptions, ...vscode.MessageItem[]],
-      Thenable<vscode.MessageItem | undefined>
-    >;
-    let telemetryStub: SinonStub<[string, string], Promise<void>>;
-    beforeEach(() => {
-      existsSyncStub = stub(fs, 'existsSync');
-      notificationStub = spy(vscode.window, 'showErrorMessage');
-      telemetryStub = stub(telemetryService, 'sendException');
-      telemetryStub.returns(Promise.resolve());
-    });
-
-    afterEach(() => {
-      existsSyncStub.restore();
-      notificationStub.restore();
-      telemetryStub.restore();
-    });
-    const root = /^win32/.test(process.platform) ? 'C:\\' : '/var';
-    const sfdxProjectPath = path.join(root, 'project', 'mockSfdxProject');
-
-    it('Should return LWC Test Runner Path when LWC Test Runner is installed and not display error message', () => {
-      existsSyncStub.returns(true);
-      const lwcTestRunnerExecutable = getLwcTestRunnerExecutable(
-        sfdxProjectPath
-      );
-      expect(lwcTestRunnerExecutable).to.equal(
-        path.join(sfdxProjectPath, 'node_modules', '.bin', 'lwc-jest')
-      );
-      assert.notCalled(notificationStub);
-      assert.notCalled(telemetryStub);
-    });
-
-    it('Should display error message when LWC Jest Test Runner is not installed', () => {
-      existsSyncStub.returns(false);
-      getLwcTestRunnerExecutable(sfdxProjectPath);
-      assert.calledOnce(notificationStub);
-      // @ts-ignore
-      assert.calledWith(
-        notificationStub,
-        nls.localize('no_lwc_jest_found_text')
-      );
-      assert.calledOnce(telemetryStub);
-      assert.calledWith(
-        telemetryStub,
-        'lwc_test_no_lwc_jest_found',
-        nls.localize('no_lwc_jest_found_text')
-      );
-    });
-  });
-
   describe('Jest Execution Info Unit Tests', () => {
     let uuidStub: SinonStub<
       [({ random: InputBuffer } | { rng(): InputBuffer } | undefined)?],
