+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/testSupport/workspace/getLwcTestRunnerExecutable.test.ts
@@ -1,0 +1,78 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import { shared as lspCommon } from '@salesforce/lightning-lsp-common';
+import { expect } from 'chai';
+import * as fs from 'fs';
+import * as path from 'path';
+import * as vscode from 'vscode';
+import { assert, SinonStub, stub } from 'sinon';
+import { nls } from '../../../../src/messages';
+import { telemetryService } from '../../../../src/telemetry';
+import {
+  getLwcTestRunnerExecutable,
+  workspaceService
+} from '../../../../src/testSupport/workspace';
+
+describe('getLwcTestRunnerExecutable Unit Tests', () => {
+  let existsSyncStub: SinonStub;
+  let notificationStub: SinonStub;
+  let telemetryStub: SinonStub;
+  let getCurrentWorkspaceTypeStub: SinonStub;
+  beforeEach(() => {
+    existsSyncStub = stub(fs, 'existsSync');
+    notificationStub = stub(vscode.window, 'showErrorMessage');
+    telemetryStub = stub(telemetryService, 'sendException');
+    getCurrentWorkspaceTypeStub = stub(
+      workspaceService,
+      'getCurrentWorkspaceType'
+    );
+    telemetryStub.returns(Promise.resolve());
+  });
+
+  afterEach(() => {
+    existsSyncStub.restore();
+    notificationStub.restore();
+    telemetryStub.restore();
+    getCurrentWorkspaceTypeStub.restore();
+  });
+  const root = /^win32/.test(process.platform) ? 'C:\\' : '/var';
+  const sfdxProjectPath = path.join(root, 'project', 'mockSfdxProject');
+
+  describe('SFDX workspace', () => {
+    beforeEach(() => {
+      getCurrentWorkspaceTypeStub.returns(lspCommon.WorkspaceType.SFDX);
+    });
+
+    it('Should return LWC Test Runner Path when LWC Test Runner is installed and not display error message', () => {
+      existsSyncStub.returns(true);
+      const lwcTestRunnerExecutable = getLwcTestRunnerExecutable(
+        sfdxProjectPath
+      );
+      expect(lwcTestRunnerExecutable).to.equal(
+        path.join(sfdxProjectPath, 'node_modules', '.bin', 'lwc-jest')
+      );
+      assert.notCalled(notificationStub);
+      assert.notCalled(telemetryStub);
+    });
+
+    it('Should display error message when LWC Jest Test Runner is not installed', () => {
+      existsSyncStub.returns(false);
+      getLwcTestRunnerExecutable(sfdxProjectPath);
+      assert.calledOnce(notificationStub);
+      assert.calledWith(
+        notificationStub,
+        nls.localize('no_lwc_jest_found_text')
+      );
+      assert.calledOnce(telemetryStub);
+      assert.calledWith(
+        telemetryStub,
+        'lwc_test_no_lwc_jest_found',
+        nls.localize('no_lwc_jest_found_text')
+      );
+    });
+  });
+});
